requirements.txt
fastapi
uvicorn
jinja2
pydantic
sqlalchemy
python-dotenv
# openai
openai==0.28